[DC-1789] Rename utils.py to notebook_utils.py
